android - enable l500 - CR fix
